refactor(vite-plugin): split dev host into structured app
