+++ packages/vite-plugin/package.json
@@ -8,7 +8,8 @@
   "types": "./dist/index.d.ts",
   "files": [
     "dist",
-    "templates"
+    "templates",
+    "src/dev-host"
   ],
   "keywords": [
     "react-native",
@@ -45,6 +46,8 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@radix-ui/react-scroll-area": "^1.2.9",
+    "@radix-ui/react-tabs": "^1.1.12",
     "@microsoft/api-extractor": "^7.50.1",
     "@vitejs/plugin-react": "^4.3.4",
     "ejs": "^3.1.9",
@@ -53,10 +56,16 @@
     "vite-plugin-react-native-web": "^2.1.1"
   },
   "peerDependencies": {
+    "react": "*",
+    "react-dom": "*",
     "vite": "catalog:"
   },
   "devDependencies": {
-    "@types/ejs": "^3.1.5"
+    "@types/ejs": "^3.1.5",
+    "@types/react": "catalog:",
+    "@types/react-dom": "catalog:",
+    "react": "catalog:",
+    "react-dom": "catalog:"
   },
   "engines": {
     "node": ">=20"
