+++ packages/vite-plugin/src/dev-host/components/ui/ScrollArea.tsx
@@ -1,0 +1,46 @@
+import {
+  forwardRef,
+  type ComponentPropsWithoutRef,
+  type ElementRef,
+  type ReactNode,
+} from 'react';
+import * as ScrollAreaPrimitive from '@radix-ui/react-scroll-area';
+import { cn } from '../../utils.js';
+
+type ScrollAreaProps = ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Root> & {
+  children: ReactNode;
+};
+
+const ScrollBar = forwardRef<
+  ElementRef<typeof ScrollAreaPrimitive.ScrollAreaScrollbar>,
+  ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.ScrollAreaScrollbar>
+>(({ className, ...props }, ref) => {
+  return (
+    <ScrollAreaPrimitive.ScrollAreaScrollbar
+      ref={ref}
+      className={cn('rz-scrollbar', className)}
+      {...props}
+    >
+      <ScrollAreaPrimitive.ScrollAreaThumb className="rz-scrollbar-thumb" />
+    </ScrollAreaPrimitive.ScrollAreaScrollbar>
+  );
+});
+
+ScrollBar.displayName = ScrollAreaPrimitive.ScrollAreaScrollbar.displayName;
+
+export const ScrollArea = forwardRef<
+  ElementRef<typeof ScrollAreaPrimitive.Root>,
+  ScrollAreaProps
+>(({ className, children, ...props }, ref) => {
+  return (
+    <ScrollAreaPrimitive.Root ref={ref} className={cn('rz-scroll-area', className)} {...props}>
+      <ScrollAreaPrimitive.Viewport className="rz-scroll-viewport">
+        {children}
+      </ScrollAreaPrimitive.Viewport>
+      <ScrollBar orientation="vertical" />
+      <ScrollAreaPrimitive.Corner />
+    </ScrollAreaPrimitive.Root>
+  );
+});
+
+ScrollArea.displayName = ScrollAreaPrimitive.Root.displayName;
