+++ packages/vite-plugin/src/dev-host/constants.ts
@@ -1,0 +1,29 @@
+export const DEV_HOST_STATE_ELEMENT_ID = '__rozenite-dev-host__';
+
+export const DEFAULT_DEVTOOLS_HEIGHT = 288;
+export const MIN_DEVTOOLS_HEIGHT = 180;
+export const MIN_IFRAME_HEIGHT = 220;
+export const DETAILS_PANEL_WIDTH = 360;
+export const MIN_DETAILS_WIDTH = 280;
+export const DEFAULT_COMMAND_WIDTH = 320;
+export const MIN_COMMAND_WIDTH = 260;
+export const SPLITTER_SIZE = 6;
+
+export const jsonTreeTheme = {
+  base00: 'transparent',
+  base01: '#374151',
+  base02: '#4b5563',
+  base03: '#6b7280',
+  base04: '#9ca3af',
+  base05: '#d1d5db',
+  base06: '#e5e7eb',
+  base07: '#f9fafb',
+  base08: '#ef4444',
+  base09: '#f59e0b',
+  base0A: '#10b981',
+  base0B: '#3b82f6',
+  base0C: '#06b6d4',
+  base0D: '#8b5cf6',
+  base0E: '#ec4899',
+  base0F: '#f97316',
+};
