+++ packages/vite-plugin/src/dev-host/dev-host.html
@@ -3,13 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title><%= pageTitle %> Dev Host</title>
+    <title>Rozenite Dev Host</title>
   </head>
   <body>
     <div id="root"></div>
-    <script>
-      window.__ROZENITE_DEV_HOST__ = <%- serializedState %>;
-    </script>
-    <script type="module" src="/@id/<%= virtualDevHostId %>"></script>
   </body>
 </html>
