+++ packages/vite-plugin/src/dev-host/types.ts
@@ -1,0 +1,33 @@
+export type DevHostPanelEntry = {
+  label: string;
+  source: string;
+};
+
+export type DevHostState = {
+  packageName: string;
+  packageDescription: string;
+  panels: DevHostPanelEntry[];
+};
+
+export type MessageEntry = {
+  id: string;
+  direction: 'in' | 'out';
+  date: string;
+  type: string;
+  payload: unknown;
+};
+
+export type PluginMessage = {
+  pluginId: string;
+  type: string;
+  payload: unknown;
+};
+
+export type ResizeHandleId = 'devtools-height' | 'command-width' | 'details-width';
+
+export type ResizeSession = {
+  handleId: ResizeHandleId;
+  pointerId: number;
+  element: HTMLElement;
+  onMove: (event: PointerEvent) => void;
+};
